Module identity_core::crypto

Cryptographic Utilities

Modules

key

merkle_tree
Types and traits for Merkle tree operations.

proof
Types and traits for helping ensure the authenticity and integrity of DID Documents and Verifiable Credentials.

signature

Structs

JcsEd25519Signature2020
An implementation of the JCS Ed25519 Signature 2020 signature suite for Linked Data Proofs.

KeyPair
A convenience for storing a pair of cryptographic keys

Signature
A DID Document digital signature.

SignatureOptions
Customizable properties of a DID Document signature.

SignatureValue
A SignatureData wrapper with a visiblity toggle.

Enums

SignatureData
A DID Document signature with a dynamic JSON field name.

Traits

SetSignature
A trait for types that can store a digital signature.

SigName
A trait for signature suites identified by a particular name.

SigSign
A trait for general-purpose signature creation

SigVerify
A trait for general-purpose signature verification

TrySignature
A trait for types that can provide a reference to a Signature.

TrySignatureMut
A trait for types that can provide a mutable reference to a Signature.

Type Definitions

PublicKey
A cryptographic key with Public components.

SecretKey
A cryptographic key with Secret components.