Module identity_core::crypto::merkle_key

An implementation of Merkle Key Collection Signatures.

Modules

base

impls

signer

traits

verifier

Structs

MerkleKey
Common utilities for working with Merkle Key Collection Signatures.

MerkleTag
A tag identifying a Merkle Key Collection signature or digest algorithm.

Sha256
The SHA-256 hash algorithm with the SHA-256 initial hash value.

Signer
A signature creation helper for Merkle Key Collection Signatures.

Verifier
A signature verification helper for Merkle Key Collection Signatures.

Traits

MerkleDigest
A common interface for digest algorithms supported by Merkle Key Signatures.

MerkleSignature
A common interface for signature algorithms supported by Merkle Key Signatures.

Type Definitions

DynSigner
An alias for a Signer with a dynamic signature type.

DynVerifier
An alias for a Verifier with a dynamic signature type.