Module identity_core::crypto::merkle_tree

Types and traits for Merkle tree operations.

Modules

consts

digest

hash

math

merkle

node

proof

tree

Structs

Hash
The output of a hash function.

MTree
A Merkle tree designed for static data.

Proof
An inclusion proof that allows proving the existence of data in a Merkle tree.

Enums

Node
A tagged Hash.

Traits

Digest
The Digest trait specifies an interface common for digest functions.

DigestExt
An extension of the Digest trait for Merkle tree construction.