Struct identity_core::crypto::merkle_tree::proof::Proof

pub struct Proof<D: DigestExt> {
    nodes: Box<[Node<D>]>,
}

An Merkle tree inclusion proof that allows proving the existence of a particular leaf in a Merkle tree.

Fields
nodes: Box<[Node<D>]>
Implementations

impl<D: DigestExt> Proof<D>

pub fn new(nodes: Box<[Node<D>]>) -> Self

Creates a new Proof from a boxed slice of nodes.

pub fn nodes(&self) -> &[Node<D>]

Returns the nodes as a slice.

pub fn index(&self) -> usize

Returns the index of underlying leaf node in the Merkle tree.

pub fn verify<T>(&self, root: &Hash<D>, target: T) -> bool where
T: AsLeaf<D>,

Verifies the computed root of self with the given root hash.

pub fn verify_hash(&self, root: &Hash<D>, hash: Hash<D>) -> bool

Verifies the computed root of self with the given root hash and a pre-computed target hash.

pub fn root(&self, target: Hash<D>) -> Hash<D>

Computes the root hash from target using a default digest.

pub fn root_with(&self, digest: &mut D, target: Hash<D>) -> Hash<D>

Computes the root hash from target using the given digest.

impl<D> Proof<D> where
D: DigestExt,

pub fn encode(&self) -> Vec<u8>

Encodes self as a vector of bytes.

See Proof::encode_into for more details.

pub fn encode_into(&self, output: &mut Vec<u8>)

Encodes self as a vector of bytes, appended to output.

The vector of bytes will have the following layout:

Node = [ U8(NODE-TAG) | NODE-HASH ]
List = [ Node0, Node1, ..., NodeN-1 ]

[ U32(LEN(List)) | List ]

pub fn decode(slice: &[u8]) -> Option<Self>

Decodes a Proof from a slice of bytes.

See Proof::encode_into for more details on the expected input format.