Module identity_core::vc

Modules

credential

credential_builder

presentation

presentation_builder

types

verifiable_credential

verifiable_presentation

Structs

Credential
A Credential represents a set of claims describing an entity.

CredentialBuilder
A CredentialBuilder is used to create a customized Credential.

CredentialSchema
Information used to validate the structure of a Credential.

CredentialStatus
Information used to determine the current status of a Credential.

CredentialSubject
An entity who is the target of a set of claims.

Evidence
Information used to increase confidence in the claims of a Credential

IssuerData
A Credential issuer in object form.

Presentation
A Presentation represents a bundle of one or more VerifiableCredentials.

PresentationBuilder
A PresentationBuilder is used to create a customized Presentation.

RefreshService
Information used to refresh or assert the status of a Credential.

TermsOfUse
Information used to express obligations, prohibitions, and permissions about a Credential or Presentation.

VerifiableCredential
A VerifiableCredential represents a Credential with an associated digital proof.

VerifiablePresentation
A VerifiablePresentation represents a Presentation with an associated digital proof.

Enums

Issuer
An identifier representing the issuer of a Credential.